Should return short for getShort method instead of returning getLong

+++ toy-common/src/main/java/org/apache/toy/ToyConfiguration.java
@@ -65,7 +65,7 @@ public class ToyConfiguration {
     return long_value.isEmpty() ? Constants.UNSET_LONG: Long.parseLong(long_value);
   }
 
-  public long getShort(String key) {
+  public short getShort(String key) {
     String short_value = get(key);
     return short_value.isEmpty() ? Constants.UNSET_SHORT: Short.parseShort(short_value);
   }
